+++ apps/src/sites/studio/pages/codeprojects_preview/page_not_found.js
@@ -1,6 +1,7 @@
-import PageNotFound from '@codebridge/FilePreview/PageNotFound';
 import React from 'react';
 import ReactDOM from 'react-dom';
+
+import PageNotFound from '@cdo/apps/weblab2/htmlPreview/PageNotFound';
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
