+++ apps/src/sites/studio/pages/codeprojects_preview/show.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import InnerHTMLPreview from '@cdo/apps/codebridge/FilePreview/InnerHTMLPreview';
+import InnerHTMLPreview from '@cdo/apps/weblab2/htmlPreview/InnerHTMLPreview';
 
 window.React = require('react');
 window.ReactDOM = require('react-dom');
