+++ apps/src/weblab2/htmlPreview/HTMLPreview.tsx
@@ -20,8 +20,7 @@ import PanelContainer from '@cdo/apps/lab2/views/components/PanelContainer';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import experiments from '@cdo/apps/util/experiments';
 import {useAppSelector, useAppDispatch} from '@cdo/apps/util/reduxHooks';
-
-import {filterSourceForPreview} from '../utils/filterSourceForPreview';
+import {filterSourceForPreview} from '@cdo/apps/weblab2/htmlPreview/filterSourceForPreview';
 
 import {
   IframeMessageType,
