+++ apps/src/weblab2/htmlPreview/contentSecurityPolicyHelper.ts
