+++ apps/src/weblab2/htmlPreview/weblab2_project_service_worker.js
