+++ dashboard/app/controllers/codeprojects_preview_controller.rb
@@ -8,7 +8,7 @@ class CodeprojectsPreviewController < ApplicationController
 
   skip_forgery_protection only: :weblab2_project_service_worker
   def weblab2_project_service_worker
-    send_file "#{apps_dir}/src/codebridge/FilePreview/weblab2_project_service_worker.js", type: 'application/javascript'
+    send_file "#{apps_dir}/src/weblab2/htmlPreview/weblab2_project_service_worker.js", type: 'application/javascript'
   end
 
   def not_found
@@ -16,7 +16,7 @@ class CodeprojectsPreviewController < ApplicationController
     render 'page_not_found', layout: false, status: :not_found
   end
 
-  # This is mostly copied over to apps/src/codebridge/FilePreview/contentSecurityPolicyHelper.ts
+  # This is mostly copied over to apps/src/weblab2/htmlPreview/contentSecurityPolicyHelper.ts
   # to generate the same Content Security Policy on the frontend. If you make changes here,
   # please make sure to update the frontend version as well. The only differences between the two are we
   # don't need to add the websocket URL in the frontend version, and we get the preview and code.org urls
